Make identifying the start of pytest commands smarter
Before if you had "pytest " in your test output we'd stop parsing.
Now we check that the prompt is in the line, and the pytest command is there which hopefully solves any issues of "pytest " existing in the test output. Should help with future output parsing too

Default prompt is based off ohmyzsh, we fallback to match everything and use the original less reliable logic