Add git status parsing
Now when you press alt and s it will read out the changes that are staged. We need to reverse the order of the lines to ensure we find the 'staged' line first, then add file names.

Currently if your buffer isn't big enough (aka it scrolled off the top of the screen) it won't work